Restructure into different modules
