+++ lib/xcvm/instructions/src/instruction.rs
@@ -1,0 +1,8 @@
+use alloc::vec::Vec;
+
+#[derive(Clone)]
+pub enum XCVMInstruction<Account, Assets> {
+	Transfer(Account, Assets),
+	Bridge(XCVMNetwork, Assets),
+	Call(Vec<u8>),
+}
