+++ lib/xcvm/instructions/src/protocol.rs
@@ -1,0 +1,6 @@
+use crate::network::XCVMNetwork;
+use alloc::vec::Vec;
+
+pub trait XCVMProtocol {
+	fn serialize(&self, network: XCVMNetwork) -> Vec<u8>;
+}
